extend test with sample data
